requirements.txt
numpy==1.26.4
cupy-cuda12x==13.4.1
scipy==1.12.0
